Added support for MySQL database

+++ db_create.py
@@ -7,12 +7,17 @@ from amalgam.models.models import User
 
 if __name__ == '__main__':
 
-    db_full_path = os.path.abspath('./amalgam.db')
-    if os.path.isfile(db_full_path):
-        print("Removing old database: {}".format(db_full_path))
-        os.remove(db_full_path)
-    else:
-        print("No database present at : {}".format(db_full_path))
+    if database.SQLALCHEMY_DATABASE == 'sqlite':
+        db_full_path = os.path.abspath('./amalgam.db')
+        if os.path.isfile(db_full_path):
+            print("Removing old database: {}".format(db_full_path))
+            os.remove(db_full_path)
+        else:
+            print("No database present at : {}".format(db_full_path))
+    elif database.SQLALCHEMY_DATABASE == 'mysql':
+        # TODO: Add a drop all table
+        #  Maybe: https://stackoverflow.com/questions/11233128/how-to-clean-the-database-dropping-all-records-using-sqlalchemy
+        pass
 
     #
     # Create all tables if needed
